Add files via upload
Changed flex powers so characters can buy armor and weapon proficiencies and increase their HP, MP, Initiative
Changed Restore Mana so it cannot be used on the caster.  I am aware that two Priests could cast it on each other, but I don't find this problematic as it would require the opportunity cost of not having other classes in the party.
Changed all available documents to simple pdfs for readability.